requirements.txt
python-telegram-bot[job-queue]==21.5
motor==3.4.0
pymongo==4.7.3
Flask==3.0.3
aiohttp==3.9.5
python-dotenv==1.0.1
gunicorn==22.0.0
